backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
supabase==1.2.0
redis==5.0.1
celery==5.3.4
langchain==0.0.350
langchain-openai==0.0.2
langgraph==0.0.26
openai==1.3.7
requests==2.31.0
feedparser==6.0.10
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
aiohttp==3.9.1
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
APScheduler==3.10.4
